app/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
psycopg2-binary==2.9.3
geopy==2.2.0
folium==0.12.1

